[build] Allow sufficient stack space for memsize_e820()
The functions memsize_e820() (and its inlined sanitize_e820_map()) use
around 9.5kb of stack space, which overflows the 8kB per-CPU stack and
so overwrites random variables in .bss, leading to undefined
behaviour.

Fix by reordering the per-CPU stacks so that the boot processor uses
the highest stack.  This provides additional space during the
start-of-day call to memsize_e820(), using stack space which will
subsequently be reused for other CPUs' stacks.

Note that we cannot simply increase the per-CPU stack size to 16kB,
since that would cause the .bss section to exceed the limits of base
memory.

Signed-off-by: Michael Brown <mcb30@ipxe.org>