[cmdline] Add "onefail" command-line parameter
The semantics of "onepass" (as implemented upstream) are such that
memtest86+ will never exit if tests fail; the failures are simply left
on screen for the user to observe while the tests continue to run.
This makes sense in the upstream build of memtest86+, where "exit"
means "reboot".  However, when running memtest86+ as a PXE NBP it
would be more useful to return to the caller after a failure (with the
appropriate exit status).

Add a command-line parameter "onefail" which will trigger an immediate
exit in the event of a failure.  Users of the PXE NBP version of
memtest86+ in an automated environment can then invoke memtest86+ with
both "onepass" and "onefail" options; for example:

  #!ipxe
  chain memtest.0 onepass onefail && goto mem_good || goto mem_bad

Signed-off-by: Michael Brown <mcb30@ipxe.org>