[test] Correct chunk calculations for test #7
Fixing the freeze of test #7 (Block Move) in SMP mode. This usually
happens in the middle of block 4096-6114M.

Function calculate_chunk(&start, &end,...) returns address of the first
(start) and the last (end) word to be tested by current cpu. The
variable end is incremented so it points just after the tested memory
(only for test #7).

After that, this big block is divided into one or more smaller blocks of
size <= 256MB. But this won't happen if integer overflow occurs during
the incrementation.

So current code may lead to different count of blocks for the last cpu
(the one with the highest memory address) and all other. Different count
of blocks means also different count of calls of function do_tick(me);
and therefore different count of calls of s_barrier();.